requirements.txt
numpy
torch
Pillow
torchvision
visdom
streamlit
